requirements.txt
pillow==9.5.0
pygments==2.14.0
